renamed czsl to clinical_zeroshot_labeler
